+++ src/main/java/mongkey/maeilmail/dto/helper/request/HelperRequestContentDto.java
@@ -1,4 +1,4 @@
-package mongkey.maeilmail.dto.helper;
+package mongkey.maeilmail.dto.helper.request;
 
 import lombok.Builder;
 import lombok.Getter;
